Revert "[MNT] rename `testing.utils.inspect` to avoid shadowing of `inspect` …"
This reverts commit 139edac6bfc732b94244a03dca9336e65c63faf7.
