+++ skbase/testing/test_all_objects.py
@@ -18,7 +18,7 @@ from skbase.lookup import all_objects
 from skbase.testing.utils._conditional_fixtures import (
     create_conditional_fixtures_and_names,
 )
-from skbase.testing.utils._inspect import _get_args
+from skbase.testing.utils.inspect import _get_args
 from skbase.utils.deep_equals import deep_equals
 from skbase.utils.dependencies import _check_soft_dependencies
