+++ skbase/tests/conftest.py
@@ -35,7 +35,7 @@ SKBASE_MODULES = (
     "skbase.testing.test_all_objects",
     "skbase.testing.utils",
     "skbase.testing.utils._conditional_fixtures",
-    "skbase.testing.utils._inspect",
+    "skbase.testing.utils.inspect",
     "skbase.testing.utils.tests",
     "skbase.testing.utils.tests.test_deep_equals",
     "skbase.tests",
@@ -67,6 +67,7 @@ SKBASE_PUBLIC_MODULES = (
     "skbase.testing",
     "skbase.testing.test_all_objects",
     "skbase.testing.utils",
+    "skbase.testing.utils.inspect",
     "skbase.testing.utils.tests",
     "skbase.testing.utils.tests.test_deep_equals",
     "skbase.tests",
@@ -203,7 +204,7 @@ SKBASE_FUNCTIONS_BY_MODULE.update(
             "_check_object_types",
             "_get_module_info",
         ),
-        "skbase.testing.utils._inspect": ("_get_args",),
+        "skbase.testing.utils.inspect": ("_get_args",),
         "skbase.utils._check": ("_is_scalar_nan",),
         "skbase.utils.dependencies": (
             "_check_soft_dependencies",
